backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
sqlalchemy==2.0.27
pydantic==2.6.1
chromadb==0.4.22
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pymupdf==1.23.21
python-docx==1.1.0
openai==1.12.0
python-dotenv==1.0.1
aiosqlite==0.19.0
greenlet==3.0.3
numpy<=2.0 